refactor(hosted-tsx): checker 与 runtime 共享同一套扫描/打包逻辑
此前 hosted-TSX 的 import/export 扫描有三份独立实现（前端 runtime 手搓扫描器、
checker 用 TS-AST、Python server 手搓 lexer），契约容易各自漂移。浏览器 runtime
无法引入完整 typescript（体积），故把唯一可共享的轻量扫描器抽成共享模块，让 checker
向 runtime 看齐。

- 新增 hostedTsxModule.mjs（纯 ESM + 手写 .d.mts）：扫描原语 + import 发现 +
  export 分类/契约 + 路径解析 + 打包，浏览器安全、零 Node/DOM 依赖
- tsxRuntime.ts 只留浏览器侧（sucrase 编译 + iframe 文档组装），bundleHostedTsxSource
  从共享模块 import（-636 行）
- check-hosted-tsx.mjs 用共享的 findHostedRelativeImportSpecifiers + assertHostedExportContract
  替换自己的 TS-AST 实现（-137 行）；ts.createProgram 仅保留做类型诊断
- 顺带补两个真实手写代码会触发的 gap：同行 ; 分隔的多 export/import（放宽语句起点判定）、
  export namespace 拒绝
- Python server 是另一门语言，不在 checker+runtime 共享范围，本次不动

模块格式选 .mjs + .d.mts（不新增 tsx 依赖）：node checker 原生 import，vue-tsc 从 .d.mts 取类型。

验证：test:hosted 56、test:hosted-script 29、type-check、eslint、
test_plugin_ui_query_service 37 全过；真实 study_companion 过检查 13 文件。
