+++ frontend/plugin-manager/src/components/plugin/hosted/tsxRuntime.ts
@@ -1,6 +1,8 @@
 import { transform } from 'sucrase'
 import type { PluginUiContext, PluginUiSurface } from '@/types/api'
 import { buildUiKitBundle } from './uiKitBundle'
+import { bundleHostedTsxSource } from './hostedTsxModule.mjs'
+import type { HostedTsxDependency } from './hostedTsxModule.mjs'
 
 type BuildHostedTsxDocumentOptions = {
   source: string
@@ -9,18 +11,6 @@ type BuildHostedTsxDocumentOptions = {
   surface: PluginUiSurface
   context?: PluginUiContext | null
   locale: string
-}
-
-type HostedTsxDependency = {
-  path: string
-  source: string
-}
-
-type HostedImportStatement = {
-  start: number
-  end: number
-  rawBindings?: string
-  specifier: string
 }
 
 function escapeScriptContent(value: string) {
@@ -37,623 +27,11 @@ function escapeHtmlAttribute(value: string) {
     .replace(/>/g, '&gt;')
 }
 
-const HOSTED_CODE_EXTENSIONS = ['.tsx', '.ts', '.jsx', '.js']
-
-function isIdentifierChar(value: string) {
-  return /[A-Za-z0-9_$]/.test(value)
-}
-
-function matchesKeyword(source: string, index: number, keyword: string) {
-  const end = index + keyword.length
-  if (source.slice(index, end) !== keyword) return false
-  const before = index > 0 ? source[index - 1] || '' : ''
-  const after = end < source.length ? source[end] || '' : ''
-  return !isIdentifierChar(before) && !isIdentifierChar(after)
-}
-
-function isHorizontalWhitespace(value: string) {
-  return value !== '\r' && value !== '\n' && /\s/.test(value)
-}
-
-function isStatementLineStart(source: string, index: number) {
-  for (let cursor = index - 1; cursor >= 0; cursor -= 1) {
-    const char = source[cursor]
-    if (char === '\n' || char === '\r') return true
-    if (!char || !isHorizontalWhitespace(char)) return false
-  }
-  return true
-}
-
-function skipWhitespace(source: string, index: number) {
-  while (index < source.length && /\s/.test(source[index] || '')) index += 1
-  return index
-}
-
-function skipHorizontalWhitespace(source: string, index: number) {
-  while (index < source.length && isHorizontalWhitespace(source[index] || '')) index += 1
-  return index
-}
-
-function skipLineComment(source: string, index: number) {
-  const newline = source.indexOf('\n', index + 2)
-  return newline < 0 ? source.length : newline + 1
-}
-
-function skipBlockComment(source: string, index: number) {
-  const end = source.indexOf('*/', index + 2)
-  return end < 0 ? source.length : end + 2
-}
-
-function skipQuoted(source: string, index: number) {
-  const quote = source[index]
-  index += 1
-  while (index < source.length) {
-    const char = source[index]
-    if (char === '\\') {
-      index += 2
-      continue
-    }
-    index += 1
-    if (char === quote) break
-  }
-  return index
-}
-
-function readQuoted(source: string, index: number) {
-  const quote = source[index]
-  index += 1
-  let value = ''
-  while (index < source.length) {
-    const char = source[index]
-    if (char === '\\') {
-      if (index + 1 < source.length) value += source[index + 1]
-      index += 2
-      continue
-    }
-    if (char === quote) return { value, end: index + 1 }
-    value += char
-    index += 1
-  }
-  return null
-}
-
-function skipTemplate(source: string, index: number) {
-  index += 1
-  while (index < source.length) {
-    const char = source[index]
-    if (char === '\\') {
-      index += 2
-      continue
-    }
-    index += 1
-    if (char === '`') break
-  }
-  return index
-}
-
-function skipTrivia(source: string, index: number) {
-  while (index < source.length) {
-    index = skipWhitespace(source, index)
-    if (index + 1 >= source.length || source[index] !== '/') return index
-    const nextChar = source[index + 1]
-    if (nextChar === '/') {
-      index = skipLineComment(source, index)
-      continue
-    }
-    if (nextChar === '*') {
-      index = skipBlockComment(source, index)
-      continue
-    }
-    return index
-  }
-  return index
-}
-
-function findKeywordBeforeStatementEnd(source: string, index: number, keyword: string) {
-  while (index < source.length) {
-    const char = source[index]
-    if (char === ';') return -1
-    if (char === '/' && index + 1 < source.length) {
-      const nextChar = source[index + 1]
-      if (nextChar === '/') {
-        index = skipLineComment(source, index)
-        continue
-      }
-      if (nextChar === '*') {
-        index = skipBlockComment(source, index)
-        continue
-      }
-    }
-    if (char === '"' || char === "'") {
-      index = skipQuoted(source, index)
-      continue
-    }
-    if (char === '`') {
-      index = skipTemplate(source, index)
-      continue
-    }
-    if (matchesKeyword(source, index, keyword)) return index
-    index += 1
-  }
-  return -1
-}
-
-function statementEndAfterSpecifier(source: string, index: number) {
-  let cursor = index
-  while (cursor < source.length) {
-    cursor = skipHorizontalWhitespace(source, cursor)
-    if (source[cursor] === '/' && source[cursor + 1] === '*') {
-      cursor = skipBlockComment(source, cursor)
-      continue
-    }
-    break
-  }
-  if (source[cursor] === ';') {
-    cursor += 1
-    while (cursor < source.length) {
-      cursor = skipHorizontalWhitespace(source, cursor)
-      if (source[cursor] === '/' && source[cursor + 1] === '*') {
-        cursor = skipBlockComment(source, cursor)
-        continue
-      }
-      break
-    }
-  }
-  if (source[cursor] === '/' && source[cursor + 1] === '/') return skipLineComment(source, cursor)
-  if (source[cursor] === '\r' && source[cursor + 1] === '\n') return cursor + 2
-  if (source[cursor] === '\r' || source[cursor] === '\n') return cursor + 1
-  return cursor
-}
-
-function readIdentifier(source: string, index: number) {
-  let cursor = index
-  if (!/[A-Za-z_$]/.test(source[cursor] || '')) return null
-  cursor += 1
-  while (cursor < source.length && isIdentifierChar(source[cursor] || '')) cursor += 1
-  return { value: source.slice(index, cursor), end: cursor }
-}
-
-function readHostedImportStatement(source: string, index: number): HostedImportStatement | null {
-  const start = index
-  index = skipTrivia(source, index + 'import'.length)
-  if (index >= source.length || source[index] === '(' || source[index] === '.') return null
-  if (source[index] === '"' || source[index] === "'") {
-    const read = readQuoted(source, index)
-    if (!read) return null
-    return { start, end: statementEndAfterSpecifier(source, read.end), specifier: read.value }
-  }
-  const bindingStart = index
-  const fromIndex = findKeywordBeforeStatementEnd(source, index, 'from')
-  if (fromIndex < 0) return null
-  const specifierIndex = skipTrivia(source, fromIndex + 'from'.length)
-  if (source[specifierIndex] !== '"' && source[specifierIndex] !== "'") return null
-  const read = readQuoted(source, specifierIndex)
-  if (!read) return null
-  return {
-    start,
-    end: statementEndAfterSpecifier(source, read.end),
-    rawBindings: source.slice(bindingStart, fromIndex).trim(),
-    specifier: read.value,
-  }
-}
-
-function hostedImportStatements(source: string) {
-  const statements: HostedImportStatement[] = []
-  let depth = 0
-  for (let index = 0; index < source.length;) {
-    const char = source[index]
-    if (char === '/' && index + 1 < source.length) {
-      const nextChar = source[index + 1]
-      if (nextChar === '/') {
-        index = skipLineComment(source, index)
-        continue
-      }
-      if (nextChar === '*') {
-        index = skipBlockComment(source, index)
-        continue
-      }
-    }
-    if (char === '"' || char === "'") {
-      index = skipQuoted(source, index)
-      continue
-    }
-    if (char === '`') {
-      index = skipTemplate(source, index)
-      continue
-    }
-    if (depth === 0 && isStatementLineStart(source, index) && matchesKeyword(source, index, 'import')) {
-      const statement = readHostedImportStatement(source, index)
-      if (statement) {
-        statements.push(statement)
-        index = statement.end
-        continue
-      }
-    }
-    if (char === '(' || char === '[' || char === '{') {
-      depth += 1
-      index += 1
-      continue
-    }
-    if (char === ')' || char === ']' || char === '}') {
-      depth = Math.max(0, depth - 1)
-      index += 1
-      continue
-    }
-    index += 1
-  }
-  return statements
-}
-
-function normalizeHostedPath(path: string) {
-  const parts: string[] = []
-  for (const segment of path.replace(/\\/g, '/').split('/')) {
-    if (!segment || segment === '.') continue
-    if (segment === '..') {
-      if (parts.length === 0) {
-        throw new Error(`Hosted TSX path escapes root: ${path}`)
-      }
-      parts.pop()
-      continue
-    }
-    parts.push(segment)
-  }
-  return parts.join('/')
-}
-
-function dirnameHostedPath(path: string) {
-  const normalized = normalizeHostedPath(path)
-  const index = normalized.lastIndexOf('/')
-  return index >= 0 ? normalized.slice(0, index) : ''
-}
-
-function resolveHostedImport(
-  fromPath: string,
-  specifier: string,
-  dependenciesByPath: Map<string, HostedTsxDependency>,
-) {
-  const cleanSpecifier = specifier.split('?', 1)[0]?.split('#', 1)[0] || ''
-  const base = normalizeHostedPath(`${dirnameHostedPath(fromPath)}/${cleanSpecifier}`)
-  const candidates = [base]
-  candidates.push(...HOSTED_CODE_EXTENSIONS.map((extension) => `${base}${extension}`))
-  candidates.push(...HOSTED_CODE_EXTENSIONS.map((extension) => `${base}/index${extension}`))
-  const resolved = candidates.find((candidate) => dependenciesByPath.has(candidate))
-  if (!resolved) {
-    throw new Error(`Missing hosted TSX dependency: ${specifier} (from ${fromPath})`)
-  }
-  return resolved
-}
-
-function parseNamedBindings(bindings: string) {
-  const trimmed = bindings.trim()
-  if (!trimmed.startsWith('{') || !trimmed.endsWith('}')) {
-    return []
-  }
-  return trimmed
-    .slice(1, -1)
-    .split(',')
-    .map((item) => item.trim())
-    .filter(Boolean)
-    .filter((item) => !isTypeOnlyBinding(item))
-    .map((item) => {
-      const aliasMatch = item.match(/^([A-Za-z_$][\w$]*)\s+as\s+([A-Za-z_$][\w$]*)$/)
-      if (aliasMatch) return `${aliasMatch[1]}: ${aliasMatch[2]}`
-      return item
-    })
-}
-
-function isTypeOnlyBinding(value: string) {
-  const trimmed = value.trim()
-  if (/^type\s*\{/.test(trimmed)) return true
-  if (/^type\s+\*\s+as\s+[A-Za-z_$][\w$]*$/.test(trimmed)) return true
-  if (/^type\s+[A-Za-z_$][\w$]*\s*,/.test(trimmed)) return true
-  return /^type\s+[A-Za-z_$][\w$]*(?:\s+as\s+[A-Za-z_$][\w$]*)?$/.test(trimmed)
-}
-
-function moduleImportStatement(rawBindings: string | undefined, modulePath: string) {
-  const bindings = String(rawBindings || '').trim()
-  const moduleRef = `__modules[${JSON.stringify(modulePath)}]`
-  if (!bindings || isTypeOnlyBinding(bindings)) {
-    return isTypeOnlyBinding(bindings) ? '' : `${moduleRef};\n`
-  }
-  const defaultNamespaceMatch = bindings.match(/^([A-Za-z_$][\w$]*)\s*,\s*\*\s+as\s+([A-Za-z_$][\w$]*)$/)
-  if (defaultNamespaceMatch?.[1] && defaultNamespaceMatch[2]) {
-    return `const ${defaultNamespaceMatch[1]} = ${moduleRef}.default;\nconst ${defaultNamespaceMatch[2]} = ${moduleRef};\n`
-  }
-  if (bindings.startsWith('* as ')) {
-    return `const ${bindings.slice(5).trim()} = ${moduleRef};\n`
-  }
-  const namedStart = bindings.indexOf('{')
-  const statements: string[] = []
-  if (namedStart > 0) {
-    const defaultName = bindings.slice(0, namedStart).trim().replace(/,$/, '').trim()
-    if (defaultName) statements.push(`const ${defaultName} = ${moduleRef}.default;`)
-  } else if (namedStart < 0) {
-    statements.push(`const ${bindings.replace(/,$/, '').trim()} = ${moduleRef}.default;`)
-  }
-  const namedBindings = namedStart >= 0 ? parseNamedBindings(bindings.slice(namedStart)) : []
-  if (namedBindings.length > 0) {
-    statements.push(`const { ${namedBindings.join(', ')} } = ${moduleRef};`)
-  }
-  return statements.length > 0 ? `${statements.join('\n')}\n` : ''
-}
-
-function hasRuntimeImportBindings(rawBindings: string | undefined) {
-  const bindings = String(rawBindings || '').trim()
-  if (!bindings) return true
-  if (isTypeOnlyBinding(bindings)) return false
-  const namedStart = bindings.indexOf('{')
-  if (namedStart < 0) return true
-  const defaultPart = bindings.slice(0, namedStart).trim().replace(/,$/, '').trim()
-  if (defaultPart) return true
-  return parseNamedBindings(bindings.slice(namedStart)).length > 0
-}
-
-function uiKitImportStatement(rawBindings: string | undefined) {
-  const bindings = String(rawBindings || '').trim()
-  const moduleRef = 'window.NekoUiKit'
-  if (!bindings || isTypeOnlyBinding(bindings)) return ''
-  const defaultNamespaceMatch = bindings.match(/^([A-Za-z_$][\w$]*)\s*,\s*\*\s+as\s+([A-Za-z_$][\w$]*)$/)
-  if (defaultNamespaceMatch?.[1] && defaultNamespaceMatch[2]) {
-    return `const ${defaultNamespaceMatch[1]} = ${moduleRef};\nconst ${defaultNamespaceMatch[2]} = ${moduleRef};\n`
-  }
-  if (bindings.startsWith('* as ')) {
-    return `const ${bindings.slice(5).trim()} = ${moduleRef};\n`
-  }
-  const namedStart = bindings.indexOf('{')
-  const statements: string[] = []
-  if (namedStart > 0) {
-    const defaultName = bindings.slice(0, namedStart).trim().replace(/,$/, '').trim()
-    if (defaultName) statements.push(`const ${defaultName} = ${moduleRef};`)
-  } else if (namedStart < 0) {
-    statements.push(`const ${bindings.replace(/,$/, '').trim()} = ${moduleRef};`)
-  }
-  const namedBindings = namedStart >= 0 ? parseNamedBindings(bindings.slice(namedStart)) : []
-  if (namedBindings.length > 0) {
-    statements.push(`const { ${namedBindings.join(', ')} } = ${moduleRef};`)
-  }
-  return statements.length > 0 ? `${statements.join('\n')}\n` : ''
-}
-
-function hostedRelativeImportPaths(
-  source: string,
-  fromPath: string,
-  dependenciesByPath: Map<string, HostedTsxDependency>,
-) {
-  const paths: string[] = []
-  for (const statement of hostedImportStatements(source)) {
-    if (statement.specifier.startsWith('./') || statement.specifier.startsWith('../')) {
-      if (!hasRuntimeImportBindings(statement.rawBindings)) continue
-      const modulePath = resolveHostedImport(fromPath, statement.specifier, dependenciesByPath)
-      paths.push(modulePath)
-    }
-  }
-  return paths
-}
-
-function orderedHostedDependencyEntries(dependenciesByPath: Map<string, HostedTsxDependency>) {
-  const ordered: Array<[string, HostedTsxDependency]> = []
-  const visited = new Set<string>()
-  const visiting: string[] = []
-
-  function visit(path: string) {
-    if (visited.has(path)) return
-    const cycleStart = visiting.indexOf(path)
-    if (cycleStart >= 0) {
-      const cycle = [...visiting.slice(cycleStart), path].join(' -> ')
-      throw new Error(`Circular hosted TSX dependency: ${cycle}`)
-    }
-    const dependency = dependenciesByPath.get(path)
-    if (!dependency) return
-
-    visiting.push(path)
-    for (const nextPath of hostedRelativeImportPaths(dependency.source, path, dependenciesByPath)) {
-      visit(nextPath)
-    }
-    visiting.pop()
-    visited.add(path)
-    ordered.push([path, dependency])
-  }
-
-  for (const path of dependenciesByPath.keys()) {
-    visit(path)
-  }
-  return ordered
-}
-
-function transformHostedImports(
-  source: string,
-  fromPath: string,
-  dependenciesByPath: Map<string, HostedTsxDependency>,
-) {
-  let result = ''
-  let cursor = 0
-  for (const statement of hostedImportStatements(source)) {
-    result += source.slice(cursor, statement.start)
-    if (statement.specifier === '@neko/plugin-ui' || statement.specifier === 'neko:ui') {
-      result += uiKitImportStatement(statement.rawBindings)
-      cursor = statement.end
-      continue
-    }
-    if (!statement.specifier.startsWith('./') && !statement.specifier.startsWith('../')) {
-      result += source.slice(statement.start, statement.end)
-      cursor = statement.end
-      continue
-    }
-    if (hasRuntimeImportBindings(statement.rawBindings)) {
-      const modulePath = resolveHostedImport(
-        fromPath,
-        statement.specifier,
-        dependenciesByPath,
-      )
-      result += moduleImportStatement(statement.rawBindings, modulePath)
-    }
-    cursor = statement.end
-  }
-  return `${result}${source.slice(cursor)}`
-}
-
-function exportAssignment(name: string, localName = name) {
-  return `__exports[${JSON.stringify(name)}] = ${localName};`
-}
-
-// A hosted module (entry or dependency) may only export with simple,
-// single-binding declaration forms: `export const/let/var NAME = ...`,
-// `export function NAME`, `export async function NAME`, `export class NAME`,
-// plus type-only declarations (`export type X`, `export interface X`) the TS
-// transform erases. The check gate (scripts/check-hosted-tsx.mjs) rejects every
-// other export form — re-exports, `export { ... }` lists, `export *`, enums,
-// generators, abstract classes, destructured/multi-declarator exports — so the
-// runtime only strips the leading `export ` keyword and records the declared
-// name. It never parses the initializer, which is what used to drag the scanner
-// into an open-ended tail of regex/template/JSX/declarator edge cases.
-const HOSTED_VALUE_EXPORT_KEYWORDS = ['function', 'class', 'const', 'let', 'var']
-
-type HostedExportDeclaration = {
-  start: number
-  declStart: number
-  name: string | null
-}
-
-function readHostedExportDeclaration(source: string, start: number): HostedExportDeclaration | null {
-  let cursor = skipTrivia(source, start + 'export'.length)
-  if (matchesKeyword(source, cursor, 'default')) return null
-  if (matchesKeyword(source, cursor, 'interface')) {
-    return { start, declStart: cursor, name: null }
-  }
-  if (matchesKeyword(source, cursor, 'type')) {
-    // `export type { ... }` / `export type * ...` are type-only export lists the
-    // gate rejects; leave them verbatim. `export type X = ...` is a declaration.
-    const afterType = skipTrivia(source, cursor + 'type'.length)
-    if (source[afterType] === '{' || source[afterType] === '*') return null
-    return { start, declStart: cursor, name: null }
-  }
-  const declStart = cursor
-  if (matchesKeyword(source, cursor, 'async')) {
-    cursor = skipTrivia(source, cursor + 'async'.length)
-  }
-  for (const keyword of HOSTED_VALUE_EXPORT_KEYWORDS) {
-    if (!matchesKeyword(source, cursor, keyword)) continue
-    const identifier = readIdentifier(source, skipTrivia(source, cursor + keyword.length))
-    return { start, declStart, name: identifier ? identifier.value : null }
-  }
-  return null
-}
-
-function hostedExportDeclarations(source: string) {
-  const statements: HostedExportDeclaration[] = []
-  let depth = 0
-  for (let index = 0; index < source.length;) {
-    const char = source[index]
-    if (char === '/' && index + 1 < source.length) {
-      const nextChar = source[index + 1]
-      if (nextChar === '/') {
-        index = skipLineComment(source, index)
-        continue
-      }
-      if (nextChar === '*') {
-        index = skipBlockComment(source, index)
-        continue
-      }
-    }
-    if (char === '"' || char === "'") {
-      index = skipQuoted(source, index)
-      continue
-    }
-    if (char === '`') {
-      index = skipTemplate(source, index)
-      continue
-    }
-    if (depth === 0 && isStatementLineStart(source, index) && matchesKeyword(source, index, 'export')) {
-      const statement = readHostedExportDeclaration(source, index)
-      if (statement) {
-        statements.push(statement)
-        index = statement.declStart
-        continue
-      }
-    }
-    if (char === '(' || char === '[' || char === '{') {
-      depth += 1
-      index += 1
-      continue
-    }
-    if (char === ')' || char === ']' || char === '}') {
-      depth = Math.max(0, depth - 1)
-      index += 1
-      continue
-    }
-    index += 1
-  }
-  return statements
-}
-
-function transformModuleExports(source: string, { handleDefault = true }: { handleDefault?: boolean } = {}) {
-  const exports: string[] = []
-  let result = ''
-  let cursor = 0
-  for (const statement of hostedExportDeclarations(source)) {
-    if (statement.start < cursor) continue
-    // Drop the leading `export ` keyword; the declaration body stays verbatim.
-    result += source.slice(cursor, statement.start)
-    if (statement.name) exports.push(exportAssignment(statement.name))
-    cursor = statement.declStart
-  }
-  let next = `${result}${source.slice(cursor)}`
-  if (handleDefault) {
-    next = next.replace(/^\s*export\s+default\s+function\s+([A-Za-z_$][\w$]*)?\s*\(/m, (_match, name) => {
-      const localName = name || '__default'
-      exports.push(exportAssignment('default', localName))
-      return `function ${localName}(`
-    })
-    next = next.replace(/^\s*export\s+default\s+/m, () => {
-      exports.push(exportAssignment('default', '__default'))
-      return 'const __default = '
-    })
-  }
-  return `${next}\n${exports.join('\n')}`
-}
-
-function sourceCommentPath(path: string) {
-  return path.replace(/\*\//g, '* /')
-}
-
-function bundleHostedTsxSource(
-  source: string,
-  dependencies: HostedTsxDependency[] = [],
-  entryPath = 'entry.tsx',
-) {
-  const dependenciesByPath = new Map<string, HostedTsxDependency>()
-  for (const dependency of dependencies) {
-    if (!dependency || typeof dependency.source !== 'string') continue
-    const normalizedPath = normalizeHostedPath(String(dependency.path || 'inline'))
-    if (dependenciesByPath.has(normalizedPath)) {
-      throw new Error(`Duplicate hosted TSX dependency path: ${normalizedPath}`)
-    }
-    dependenciesByPath.set(normalizedPath, dependency)
-  }
-  const chunks = orderedHostedDependencyEntries(dependenciesByPath)
-    .map(([path, dependency]) => {
-      const moduleSource = transformModuleExports(
-        transformHostedImports(dependency.source, path, dependenciesByPath),
-      )
-      return `
-/* hosted dependency: ${sourceCommentPath(path)} */
-__modules[${JSON.stringify(path)}] = (() => {
-  const __exports = {};
-${moduleSource}
-  return __exports;
-})();`
-    })
-  const normalizedEntryPath = normalizeHostedPath(entryPath)
-  // Transform the entry's NAMED exports too (e.g. `export const title = ...`),
-  // which would otherwise survive into the classic <script> as a syntax error;
-  // keep `export default` for compileHostedTsx to turn into __Panel.
-  const entrySource = transformModuleExports(
-    transformHostedImports(source, normalizedEntryPath, dependenciesByPath),
-    { handleDefault: false },
-  )
-  return `const __modules = Object.create(null);\n${chunks.join('\n')}\nconst __exports = {};\n${entrySource}`
-}
-
+// Bundling, dependency linking, and the import/export contract live in the
+// shared hostedTsxModule.mjs so this runtime and the check gate
+// (scripts/check-hosted-tsx.mjs) never drift. This file only owns the
+// browser-specific steps: compiling the bundled source with sucrase and
+// assembling the sandboxed iframe document.
 function compileHostedTsx(source: string, dependencies: HostedTsxDependency[] = [], entryPath = 'entry.tsx') {
   const compiled = transform(bundleHostedTsxSource(source, dependencies, entryPath), {
     transforms: ['typescript', 'jsx'],
